Fix Qwen3.5-397B-A17B-FP8 serving on Gaudi3
1. hpu_gated_deltanet.py: Cast state_indices to .long() for index_copy_()
   which requires a LongTensor index.

2. ops.py: WAR for FusedSDPA complex_guid_extractor error 400 when
   head_dim > 128. Use is_causal=True with valid_seq_lengths instead of
   explicit attn_bias, letting Synapse handle causal masking internally.

3. utils.py: Fix async_h2d_update for graph mode by replacing advanced
   indexing (synStatus 30) with index_copy_ using a device tensor.

4. hpu_model_runner.py: Fix IndexError in get_habana_paged_attn_buffers
   by ensuring block_bucket_size >= actual_blocks_needed when contiguous
   PA is enabled and fallback bucketing caps the context size.

+++ vllm_gaudi/ops/hpu_gated_deltanet.py
@@ -192,7 +192,8 @@ class HPUQwen3_5GatedDeltaNet(Qwen3_5GatedDeltaNet):
             # Write back cache
             assert final_state is not None
 
-            ssm_state.index_copy_(0, state_indices, final_state.to(device=ssm_state.device, dtype=ssm_state.dtype))
+            ssm_state.index_copy_(0, state_indices.long(), final_state.to(device=ssm_state.device,
+                                                                          dtype=ssm_state.dtype))
 
             # Copy output
             non_spec_out = core_attn_out_result.squeeze(0)
